requirements.txt
tkinter
numpy
pycryptodome
pymysql
scikit-learn
json